--- src/main/java/nl/ou/debm/common/Expr.g4
@@ -1,9 +1,0 @@
-grammar Expr;		
-prog:	expr EOF ;
-expr:	expr ('*'|'/') expr
-    |	expr ('+'|'-') expr
-    |	INT
-    |	'(' expr ')'
-    ;
-NEWLINE : [\r\n]+ -> skip;
-INT     : [0-9]+ ;
