--- src/main/java/nl/ou/debm/common/ExprListener.java
@@ -1,29 +1,0 @@
-// Generated from Expr.g by ANTLR 4.7.2
-import org.antlr.v4.runtime.tree.ParseTreeListener;
-
-/**
- * This interface defines a complete listener for a parse tree produced by
- * {@link ExprParser}.
- */
-public interface ExprListener extends ParseTreeListener {
-	/**
-	 * Enter a parse tree produced by {@link ExprParser#prog}.
-	 * @param ctx the parse tree
-	 */
-	void enterProg(ExprParser.ProgContext ctx);
-	/**
-	 * Exit a parse tree produced by {@link ExprParser#prog}.
-	 * @param ctx the parse tree
-	 */
-	void exitProg(ExprParser.ProgContext ctx);
-	/**
-	 * Enter a parse tree produced by {@link ExprParser#expr}.
-	 * @param ctx the parse tree
-	 */
-	void enterExpr(ExprParser.ExprContext ctx);
-	/**
-	 * Exit a parse tree produced by {@link ExprParser#expr}.
-	 * @param ctx the parse tree
-	 */
-	void exitExpr(ExprParser.ExprContext ctx);
-}
