--- src/main/java/nl/ou/debm/common/lib.java
@@ -1,11 +1,0 @@
-package nl.ou.debm.common;
-
-
-/**
- * This is test code
- */
-public class lib {
-    public static int inc(int i) {
-        return i+1;
-    }
-}
